Player Trading Shop + Multiple Item Selection
Added the player trading shop, and made it where the GM can select
multiple items in the Shop to buy or sell, and players can trade
multiple items at the same time with other players. Still contains all
the checks, also optimized and organized some of the code in the player
trading shop and shop.